Màj des entités et les services et controller associés

+++ src/main/java/springboot/projetfinal/repo/AuthentificationRepository.java
@@ -3,5 +3,5 @@ package springboot.projetfinal.repo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import springboot.projetfinal.model.Authentification;
 
-public interface AuthentificationRepository extends JpaRepository<Authentification, Long> {
+public interface AuthentificationRepository extends JpaRepository<Authentification, Integer> {
 }
